+++ .eslintrc.js
@@ -1,5 +1,9 @@
 module.exports = {
-  'extends': 'ryansobol/es5',
+  'extends': [
+    'ryansobol/es5',
+    'ryansobol/node',
+  ],
+
   'rules': {
     // For consistency with keys that require quotes
     'quote-props': 0