+++ es5/index.js
@@ -1,11 +1,5 @@
 'use strict';
 
 module.exports = {
-  extends: 'ryansobol/core',
-
-  env: {
-    node: true
-  },
-
-  rules: require('../rules/node')
+  extends: 'ryansobol/core'
 };