+++ es6/index.js
@@ -2,16 +2,16 @@
 
 module.exports = {
   extends: 'ryansobol/core',
-  parser: 'babel-eslint',
-
-  env: {
-    es6: true,
-    node: true
-  },
 
   ecmaFeatures: {
     modules: true
   },
+
+  env: {
+    es6: true
+  },
+
+  parser: 'babel-eslint',
 
   rules: require('../rules/ecmascript-6')
 };