+++ node/index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+module.exports = {
+  env: {
+    node: true
+  },
+
+  rules: require('../rules/node')
+};